Cleaned up code to follow industry standards
