+++ backend/src/main/java/com/apostoli/UnluckyApp/model/enums/RoleType.java
@@ -4,5 +4,5 @@ public enum RoleType {
     USER,
     RESPONDER, //rules over users
     ADMIN, //rules over responders and users
-    SUPER_USER //rules over everyone
+    SUPER_ADMIN //rules over everyone
 }
