+++ backend/src/main/java/com/apostoli/UnluckyApp/service/CityService.java
@@ -1,57 +1,9 @@
 package com.apostoli.UnluckyApp.service;
 
-
 import com.apostoli.UnluckyApp.model.entity.City;
-import com.apostoli.UnluckyApp.repository.CityRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.core.io.ClassPathResource;
-import org.springframework.stereotype.Service;
 
-import java.io.BufferedReader;
-import java.io.InputStreamReader;
-import java.nio.charset.StandardCharsets;
-import java.util.ArrayList;
 import java.util.List;
 
-@Service
-public class CityService {
-
-    private final CityRepository cityRepository;
-    private List<City> cities = new ArrayList<>();
-
-
-    @Autowired
-    public CityService(CityRepository cityRepository) {
-        this.cityRepository = cityRepository;
-        loadCities();
-    }
-
-    private void loadCities() {
-        try{
-            ClassPathResource resource = new ClassPathResource("CroatianCities.csv");
-            BufferedReader reader = new BufferedReader(new InputStreamReader(resource.getInputStream(), StandardCharsets.UTF_8));
-            String line;
-
-            reader.readLine(); // skip the header
-
-            while ((line = reader.readLine()) != null) {
-                String[] parts = line.split(",");
-                String name = parts[0];
-                double latitude = Double.parseDouble(parts[2]);
-                double longitude = Double.parseDouble(parts[3]);
-                City city = new City(name, latitude, longitude);
-                cities.add(city);
-                cityRepository.save(city);
-            }
-
-        }catch (Exception e){
-            e.printStackTrace();
-        }
-    }
-
-    public List<City> getAllCities() {
-        return cities;
-    }
-
-
+public interface CityService {
+    List<City> getAllCities();
 }
