+++ backend/src/main/java/com/apostoli/UnluckyApp/service/impl/PhotoServiceImpl.java
@@ -1,0 +1,20 @@
+package com.apostoli.UnluckyApp.service.impl;
+
+
+import com.apostoli.UnluckyApp.repository.ReportRepository;
+import com.apostoli.UnluckyApp.service.PhotoService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+
+@Service
+public class PhotoServiceImpl implements PhotoService {
+
+
+    private final ReportRepository reportRepository;
+
+    @Autowired
+    public PhotoServiceImpl(ReportRepository reportRepository) {
+        this.reportRepository = reportRepository;
+    }
+}
